This function gets the currently assigned function pointer assignments

This function gets the currently assigned function pointer assignments

rxode2parseGetPointerAssignment()

Value

The currently assigned pointer assignments

Author

Matthew L. Fidler

Examples

rxode2parseGetTranslation()
#>                  rxFun                  fun                  type      package
#> 1              linCmtA              linCmtA             linCmtA_p  rxode2parse
#> 2              linCmtB              linCmtB             linCmtB_p  rxode2parse
#> 3              linCmtC              linCmtC             linCmtA_p  rxode2parse
#> 4               rxnorm               rxnorm           rxode2i_fn2 rxode2random
#> 5              rxbinom              rxbinom       rxode2i_rxbinom rxode2random
#> 6             rxnbinom             rxnbinom       rxode2i_rxbinom rxode2random
#> 7           rxnbinomMu           rxnbinomMu       rxode2i_rxbinom rxode2random
#> 8             rxcauchy             rxcauchy           rxode2i_fn2 rxode2random
#> 9              rxchisq              rxchisq            rxode2i_fn rxode2random
#> 10               rxexp                rxexp            rxode2i_fn rxode2random
#> 11                 rxf                  rxf           rxode2i_fn2 rxode2random
#> 12              rxgeom               rxgeom           rxode2i_ifn rxode2random
#> 13             rxgamma              rxgamma           rxode2i_fn2 rxode2random
#> 14              rxbeta               rxbeta           rxode2i_fn2 rxode2random
#> 15              rxpois               rxpois           rxode2i_ifn rxode2random
#> 16                rxt_                 rxt_            rxode2i_fn rxode2random
#> 17              rxunif               rxunif           rxode2i_fn2 rxode2random
#> 18           rxweibull            rxweibull           rxode2i_fn2 rxode2random
#> 19              rinorm               rinorm          rxode2i2_fn2 rxode2random
#> 20             ribinom              ribinom      rxode2i2_ribinom rxode2random
#> 21            rinbinom             rinbinom      rxode2i2_ribinom rxode2random
#> 22          rinbinomMu           rinbinomMu      rxode2i2_ribinom rxode2random
#> 23            ricauchy             ricauchy          rxode2i2_fn2 rxode2random
#> 24             richisq              richisq           rxode2i2_fn rxode2random
#> 25               riexp                riexp           rxode2i2_fn rxode2random
#> 26                 rif                  rif          rxode2i2_fn2 rxode2random
#> 27              rigeom               rigeom          rxode2i2_ifn rxode2random
#> 28             rigamma              rigamma          rxode2i2_fn2 rxode2random
#> 29              ribeta               ribeta          rxode2i2_fn2 rxode2random
#> 30              ripois               ripois          rxode2i2_ifn rxode2random
#> 31                rit_                 rit_           rxode2i2_fn rxode2random
#> 32              riunif               riunif          rxode2i2_fn2 rxode2random
#> 33           riweibull            riweibull          rxode2i2_fn2 rxode2random
#> 34                 phi                  phi             rxode2_fn rxode2random
#> 35              gammap               gammap            rxode2_fn2       rxode2
#> 36              gammaq               gammaq            rxode2_fn2       rxode2
#> 37           gammapInv            gammapInv            rxode2_fn2       rxode2
#> 38          gammapInva           gammapInva            rxode2_fn2       rxode2
#> 39           gammaqInv            gammaqInv            rxode2_fn2       rxode2
#> 40          gammaqInva           gammaqInva            rxode2_fn2       rxode2
#> 41          uppergamma           uppergamma            rxode2_fn2       rxode2
#> 42          lowergamma           lowergamma            rxode2_fn2       rxode2
#> 43           gammapDer            gammapDer            rxode2_fn2       rxode2
#> 44               logit                logit            rxode2_fn3       rxode2
#> 45               expit                expit            rxode2_fn3       rxode2
#> 46              simeta               simeta               _simfun rxode2random
#> 47              simeps               simeps               _simfun rxode2random
#> 48            llikNorm            _llikNorm    rxode2_llikNormFun     rxode2ll
#> 49       llikNormDmean       _llikNormDmean    rxode2_llikNormFun     rxode2ll
#> 50         llikNormDsd         _llikNormDsd    rxode2_llikNormFun     rxode2ll
#> 51            llikPois            _llikPois    rxode2_llikPoisFun     rxode2ll
#> 52     llikPoisDlambda     _llikPoisDlambda    rxode2_llikPoisFun     rxode2ll
#> 53           llikBinom           _llikBinom   rxode2_llikBinomFun     rxode2ll
#> 54      llikBinomDprob      _llikBinomDprob   rxode2_llikBinomFun     rxode2ll
#> 55          llikNbinom          _llikNbinom   rxode2_llikBinomFun     rxode2ll
#> 56     llikNbinomDprob     _llikNbinomDprob   rxode2_llikBinomFun     rxode2ll
#> 57        llikNbinomMu        _llikNbinomMu   rxode2_llikBinomFun     rxode2ll
#> 58     llikNbinomMuDmu     _llikNbinomMuDmu   rxode2_llikBinomFun     rxode2ll
#> 59            llikBeta            _llikBeta    rxode2_llikBetaFun     rxode2ll
#> 60     llikBetaDshape1     _llikBetaDshape1    rxode2_llikBetaFun     rxode2ll
#> 61     llikBetaDshape2     _llikBetaDshape2    rxode2_llikBetaFun     rxode2ll
#> 62               llikT               _llikT       rxode2_llikTFun     rxode2ll
#> 63            llikTDdf            _llikTDdf       rxode2_llikTFun     rxode2ll
#> 64          llikTDmean          _llikTDmean       rxode2_llikTFun     rxode2ll
#> 65            llikTDsd            _llikTDsd       rxode2_llikTFun     rxode2ll
#> 66           llikChisq           _llikChisq   rxode2_llikChisqFun     rxode2ll
#> 67        llikChisqDdf        _llikChisqDdf   rxode2_llikChisqFun     rxode2ll
#> 68             llikExp             _llikExp     rxode2_llikExpFun     rxode2ll
#> 69        llikExpDrate        _llikExpDrate     rxode2_llikExpFun     rxode2ll
#> 70               llikF               _llikF       rxode2_llikFFun     rxode2ll
#> 71           llikFDdf1           _llikFDdf1       rxode2_llikFFun     rxode2ll
#> 72           llikFDdf2           _llikFDdf2       rxode2_llikFFun     rxode2ll
#> 73            llikGeom            _llikGeom    rxode2_llikGeomFun     rxode2ll
#> 74          llikGeomDp          _llikGeomDp    rxode2_llikGeomFun     rxode2ll
#> 75            llikUnif            _llikUnif    rxode2_llikUnifFun     rxode2ll
#> 76      llikUnifDalpha      _llikUnifDalpha    rxode2_llikUnifFun     rxode2ll
#> 77       llikUnifDbeta       _llikUnifDbeta    rxode2_llikUnifFun     rxode2ll
#> 78         llikWeibull         _llikWeibull rxode2_llikWeibullFun     rxode2ll
#> 79   llikWeibullDshape   _llikWeibullDshape rxode2_llikWeibullFun     rxode2ll
#> 80   llikWeibullDscale   _llikWeibullDscale rxode2_llikWeibullFun     rxode2ll
#> 81           llikGamma           _llikGamma   rxode2_llikGammaFun     rxode2ll
#> 82     llikGammaDshape     _llikGammaDshape   rxode2_llikGammaFun     rxode2ll
#> 83      llikGammaDrate      _llikGammaDrate   rxode2_llikGammaFun     rxode2ll
#> 84          llikCauchy          _llikCauchy  rxode2_llikCauchyFun     rxode2ll
#> 85 llikCauchyDlocation _llikCauchyDlocation  rxode2_llikCauchyFun     rxode2ll
#> 86    llikCauchyDscale    _llikCauchyDscale  rxode2_llikCauchyFun     rxode2ll
#>               packageFun argMin argMax threadSafe
#> 1                linCmtA     20     20          1
#> 2                linCmtB     21     21          1
#> 3                linCmtC     20     20          1
#> 4                 rxnorm     NA     NA          1
#> 5                rxbinom     NA     NA          1
#> 6               rxnbinom     NA     NA          1
#> 7             rxnbinomMu     NA     NA          1
#> 8               rxcauchy     NA     NA          1
#> 9                rxchisq     NA     NA          1
#> 10                 rxexp     NA     NA          1
#> 11                   rxf     NA     NA          1
#> 12                rxgeom     NA     NA          1
#> 13               rxgamma     NA     NA          1
#> 14                rxbeta     NA     NA          1
#> 15                rxpois     NA     NA          1
#> 16                  rxt_     NA     NA          1
#> 17                rxunif     NA     NA          1
#> 18             rxweibull     NA     NA          1
#> 19                rinorm     NA     NA          1
#> 20               ribinom     NA     NA          1
#> 21              rinbinom     NA     NA          1
#> 22            rinbinomMu     NA     NA          1
#> 23              ricauchy     NA     NA          1
#> 24               richisq     NA     NA          1
#> 25                 riexp     NA     NA          1
#> 26                   rif     NA     NA          1
#> 27                rigeom     NA     NA          1
#> 28               rigamma     NA     NA          1
#> 29                ribeta     NA     NA          1
#> 30                ripois     NA     NA          1
#> 31                  rit_     NA     NA          1
#> 32                riunif     NA     NA          1
#> 33             riweibull     NA     NA          1
#> 34                   phi     NA     NA          1
#> 35                gammap      2      2          1
#> 36                gammaq      2      2          1
#> 37             gammapInv      2      2          1
#> 38            gammapInva      2      2          1
#> 39             gammaqInv      2      2          1
#> 40            gammaqInva      2      2          1
#> 41            uppergamma      2      2          1
#> 42            lowergamma      2      2          1
#> 43             gammapDer      2      2          1
#> 44                 logit     NA     NA          1
#> 45                 expit     NA     NA          1
#> 46                simeta     NA     NA          1
#> 47                simeps     NA     NA          1
#> 48            rxLlikNorm      3      3          1
#> 49       rxLlikNormDmean      3      3          1
#> 50         rxLlikNormDsd      3      3          1
#> 51            rxLlikPois      2      2          1
#> 52     rxLlikPoisDlambda      2      2          1
#> 53           rxLlikBinom      3      3          1
#> 54      rxLlikBinomDprob      3      3          1
#> 55          rxLlikNbinom      3      3          1
#> 56     rxLlikNbinomDprob      3      3          1
#> 57        rxLlikNbinomMu      3      3          1
#> 58     rxLlikNbinomMuDmu      3      3          1
#> 59            rxLlikBeta      3      3          1
#> 60     rxLlikBetaDshape1      3      3          1
#> 61     rxLlikBetaDshape2      3      3          1
#> 62               rxLlikT      4      4          1
#> 63            rxLlikTDdf      4      4          1
#> 64          rxLlikTDmean      4      4          1
#> 65            rxLlikTDsd      4      4          1
#> 66           rxLlikChisq      2      2          1
#> 67        rxLlikChisqDdf      2      2          1
#> 68             rxLlikExp      2      2          1
#> 69        rxLlikExpDrate      2      2          1
#> 70               rxLlikF      3      3          1
#> 71           rxLlikFDdf1      3      3          1
#> 72           rxLlikFDdf2      3      3          1
#> 73            rxLlikGeom      2      2          1
#> 74          rxLlikGeomDp     NA     NA          1
#> 75            rxLlikUnif      3      3          1
#> 76      rxLlikUnifDalpha      3      3          1
#> 77       rxLlikUnifDbeta      3      3          1
#> 78         rxLlikWeibull      3      3          1
#> 79   rxLlikWeibullDshape      3      3          1
#> 80   rxLlikWeibullDscale      3      3          1
#> 81           rxLlikGamma      3      3          1
#> 82     rxLlikGammaDshape      3      3          1
#> 83      rxLlikGammaDrate      3      3          1
#> 84          rxLlikCauchy      3      3          1
#> 85 rxLlikCauchyDlocation      3      3          1
#> 86    rxLlikCauchyDscale      3      3          1